# numguess-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  contact:
    name: Loyola University Chicago COMP 373/473 Software Architecture Course
    url: https://github.com/lucswarchcourse/numguess-openapi
  description: |
    A RESTful API for a number guessing game.

    The game allows players to:
    - Create new game instances
    - Submit guesses for a number between 1 and 100
    - Track the number of guesses and best scores

    ## Richardson Maturity Model Level 3 (HATEOAS)

    This API implements **Level 3 of the Richardson Maturity Model** by providing comprehensive
    hypermedia controls in all JSON responses. Clients can navigate the entire API by following
    links, without hardcoding URLs or business logic about available actions.

    ### Hypermedia Format

    The API uses **HAL (Hypertext Application Language)** conventions with `_links` objects
    containing link relations. Supported media types:
    - `application/json` - JSON with embedded hypermedia controls
    - `application/hal+json` - Explicit HAL format
    - `application/xhtml+xml` - HTML forms (for browser interaction)

    ### Standard Link Relations

    All responses include a `_links` object with the following standard relations:

    - **self**: Link to the current resource
    - **games**: Link to the games collection
    - **create-game**: Action link to create a new game (POST)
    - **submit-guess**: Action link to submit a guess (POST) - only present when game is active
...
    - **root**: Link back to API root

...

...

...

...

...

...
servers:
...
paths:
...
  /games:
    get:
      description: |
        Returns a welcome page (HTML) or games collection representation with hypermedia controls (JSON).
        For HTML clients, this displays a form to start a new game.
        For JSON clients, provides discoverable actions and navigation links (Level 3 HATEOAS).
      parameters:
      - description: Content type preference
        explode: false
        in: header
        name: Accept
        required: false
        schema:
          default: application/xhtml+xml
          enum:
          - application/xhtml+xml
          - application/json
          - application/hal+json
          type: string
        style: simple
      responses:
        "200":
          content:
            application/xhtml+xml:
              example: |
                <html>
                  <head><title>Number Guessing Game: Welcome</title></head>
                  <body>
                    <form method="post" action="">
                      <div>
                        <input type="submit" value="play" />
                      </div>
                    </form>
                  </body>
                </html>
              schema:
                format: html
                type: string
            application/json:
              example:
                message: Welcome to the Number Guessing Game! Create a new game to
                  start playing.
                totalGames: 5
                _links:
                  self:
                    href: http://localhost:3000/numguess-restlet/games
                    type: application/json
                  create-game:
                    href: http://localhost:3000/numguess-restlet/games
                    method: POST
                    type: application/json
                    title: Create a new game
                  root:
                    href: http://localhost:3000/numguess-restlet/
                    type: application/json
                    title: API root
              schema:
                $ref: "#/components/schemas/GamesCollection"
            application/hal+json:
              schema:
                $ref: "#/components/schemas/GamesCollection"
          description: Welcome page or games collection representation with hypermedia
            controls
      summary: Get welcome page or games collection
      tags:
      - games
      x-accepts:
      - application/hal+json
      - application/json
      - application/xhtml+xml
      x-tags:
      - tag: games
    post:
      description: |
        Creates a new number guessing game instance and returns hypermedia controls for interaction.
        The game is initialized with a random number between 1 and 100.
        A UUID is generated to uniquely identify the new game instance.

        Following Level 3 HATEOAS principles, the response includes comprehensive hypermedia links
        for all available actions: submitting guesses, deleting the game, or creating another game.
      parameters:
      - description: Content type preference
        explode: false
        in: header
        name: Accept
        required: false
        schema:
          default: application/xhtml+xml
          enum:
          - application/xhtml+xml
          - application/json
          - application/hal+json
          type: string
        style: simple
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
        description: Empty body or form data (optional)
        required: false
      responses:
        "201":
          content:
            application/xhtml+xml:
              schema:
                format: html
                type: string
            application/json:
              example:
                gameId: 550e8400-e29b-41d4-a716-446655440000
                message: Game created successfully. Submit your first guess!
                _links:
                  self:
                    href: http://localhost:3000/numguess-restlet/games/550e8400-e29b-41d4-a716-446655440000
                    type: application/json
                  submit-guess:
                    href: http://localhost:3000/numguess-restlet/games/550e8400-e29b-41d4-a716-446655440000
                    method: POST
                    type: application/x-www-form-urlencoded
                    title: Submit a guess
                  delete:
                    href: http://localhost:3000/numguess-restlet/games/550e8400-e29b-41d4-a716-446655440000
                    method: DELETE
                    title: Delete this game
                  games:
                    href: http://localhost:3000/numguess-restlet/games
                    type: application/json
                    title: Games collection
              schema:
                $ref: "#/components/schemas/GameCreationResponse"
            application/hal+json:
              schema:
                $ref: "#/components/schemas/GameCreationResponse"
          description: Game created successfully with hypermedia controls
          headers:
            Location:
              description: URI of the newly created game resource
              explode: false
              schema:
                example: http://localhost:3000/numguess-restlet/games/550e8400-e29b-41d4-a716-446655440000
                format: uri
                type: string
              style: simple
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal server error
      summary: Create a new game
      tags:
      - games
      x-content-type: application/x-www-form-urlencoded
      x-accepts:
      - application/hal+json
      - application/json
      - application/xhtml+xml
      x-tags:
      - tag: games
...
components:
...
  schemas:
    Link:
      description: A hypermedia link following HAL conventions
      example:
        templated: false
        method: GET
        href: https://openapi-generator.tech
        type: type
        title: title
      properties:
        href:
          description: The URI of the linked resource
          format: uri
          type: string
        templated:
          default: false
          description: Whether the href is a URI template
          type: boolean
        type:
          description: Media type hint for the target resource
          type: string
        method:
          description: HTTP method to use (for action links)
          enum:
          - GET
          - POST
          - PUT
          - PATCH
          - DELETE
          type: string
        title:
          description: Human-readable title for the link
          type: string
      required:
      - href
      type: object
...
    GamesCollection:
      description: Games collection resource with hypermedia controls
      properties:
        message:
          description: Welcome or informational message
          example: Welcome to the Number Guessing Game! Create a new game to start
            playing.
          type: string
        totalGames:
          description: Total number of active games (optional)
          example: 5
          type: integer
        _links:
          $ref: "#/components/schemas/GamesCollection__links"
      required:
      - _links
      type: object
    GameCreationResponse:
      description: Response containing hypermedia controls for newly created game
        (HATEOAS Level 3)
      properties:
        gameId:
          description: The unique identifier of the newly created game
          example: 550e8400-e29b-41d4-a716-446655440000
          format: uuid
          type: string
        href:
          deprecated: true
          description: URI of the newly created game resource (deprecated - use _links.self
            instead)
          example: http://localhost:3000/numguess-restlet/games/550e8400-e29b-41d4-a716-446655440000
          format: uri
          type: string
        message:
          description: Success message
          example: Game created successfully. Submit your first guess!
          type: string
        _links:
          $ref: "#/components/schemas/GameCreationResponse__links"
      required:
      - _links
      - gameId
      type: object
...
    Error:
      description: Error response
      example:
        error: error
        status: 0
      properties:
        error:
          description: Error message
          type: string
        status:
          description: HTTP status code
          type: integer
      type: object
...
    GamesCollection__links:
      properties:
        self:
          $ref: "#/components/schemas/Link"
        create-game:
          allOf:
          - $ref: "#/components/schemas/Link"
          description: Link to create a new game
        root:
          allOf:
          - $ref: "#/components/schemas/Link"
          description: Link back to API root
      required:
      - create-game
      - self
      type: object
    GameCreationResponse__links:
      properties:
        self:
          allOf:
          - $ref: "#/components/schemas/Link"
          description: Link to this game resource
        submit-guess:
          allOf:
          - $ref: "#/components/schemas/Link"
          description: Action link to submit a guess
        delete:
          allOf:
          - $ref: "#/components/schemas/Link"
          description: Action link to delete this game
        games:
          allOf:
          - $ref: "#/components/schemas/Link"
          description: Link back to games collection
      required:
      - delete
      - games
      - self
      - submit-guess
      type: object
...